requirements.txt
rich>=13.0.0
python-dateutil>=2.8.2
pytest>=7.0.0
